- remove hook usage - rename useStopGroup method as simple util method getStopGroup - use useMemo to cache result of getStopGroup to prevent infinite rendering
